% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetEpiDataNew.R
\name{GetEpiDataNew}
\alias{GetEpiDataNew}
\title{GetEpiDataNew function
This function read the running epi data online}
\usage{
GetEpiDataNew()
}
\description{
GetEpiDataNew function
This function read the running epi data online
}
